Seperated delete, and create crud functions to a helper file
