+++ Notes_App_Front/src/components/modules/NoteItem/index.js
@@ -1,13 +1,22 @@
 import React from "react"
 import "./index.scss"
 
+import CrudHelpers from "../../../helpers/CrudHelpers.js"
+
 const NoteItem = (props) => {
+  const { note, setNotes, notes } = props;
+
+  const handleClick = (e) => {
+    CrudHelpers.deleteNote(e, note._id)
+    setNotes(notes.filter(({ _id: i }) => note._id !== i))
+  }
+  
   return (
     <li
       key={props.note._id}
     >
       {props.note.title}
-      <button onClick={e => props.deleteNote(e, props.note._id)}>X</button>
+      <button onClick={e => handleClick(e)}>X</button>
     </li>
   )
 }
